fix(#2868): move the portfolio panels with the split, and trust the backend's block rule
Three defects Codex found in the first cut, all real:

The status verdict re-derived "can entries happen" from `execution_enabled`
alone. The backend owns that rule — `StrategyEntryBlockState.new_entries_blocked`
is kill OR any execution block OR auto-trading off — and `execution_block_reasons`
carries blocks raised by strategy_execution_blocks independently of both other
terms. The page could therefore report "Trading" while the backend refused every
entry. It now surfaces the backend's own reasons, maps the three known ones to
plain English, passes unmapped reasons through verbatim, and fails closed when
entries are blocked for a reason the payload does not enumerate.

The split was cosmetic: StrategiesPage still fetched and rendered portfolio
performance, the P&L chart, account evidence and the pot's funding control, so
the research lens re-rendered the whole portfolio dataset and its broker-facing
controls underneath the candidate list. Those panels moved to
components/strategies/StrategyPortfolioPanels.tsx and are wired into the
portfolio lens; the research lens no longer fetches pnl history or owned
positions at all.

Blocker rows keyed on `blocker.key`, which is not unique — `entries_blocked` is
emitted once per backend reason, so React could reuse or drop a row. Keyed on
kind plus label.

Two more from the second pass: closing a position no longer refreshed the P&L
chart (it moved onto this lens, so this callback owns it), and the open-positions
tile read `positions.length`, showing 0 while that request was pending or failed
— a false number sitting next to its own error. It reads the overview's
`active_position_count` sum, which is already loaded.

Tests moved with the panels rather than being deleted: they now render the hub at
the lens under test, which is what the operator reaches. Two that spanned both
lenses were split rather than trimmed. Verified on the live dev stack — portfolio
2,239 chars, research 4,623, no duplication, zero console errors.

Refs #2868
